Avoid printing *expected* stack traces during test runs
