+++ hornetq-core/src/test/java/org/hornetq/util/XMLUtilTest.java
@@ -14,8 +14,8 @@
 package org.hornetq.util;
 
 import junit.framework.Assert;
-import junit.framework.TestCase;
 
+import org.hornetq.tests.util.SilentTestCase;
 import org.hornetq.utils.XMLUtil;
 import org.w3c.dom.Element;
 import org.w3c.dom.Node;
@@ -25,7 +25,7 @@ import org.w3c.dom.NodeList;
  * @author <a href="mailto:ovidiu@feodorov.com">Ovidiu Feodorov</a>
  * @version <tt>$Revision$</tt>
  */
-public class XMLUtilTest extends TestCase
+public class XMLUtilTest extends SilentTestCase
 {
    // Constructors --------------------------------------------------
